requirements.txt
py-solc-x
web3
scikit-learn
numpy
pandas
flask
flask-cors
requests
paho-mqtt
python-dotenv
